+++ app/Http/Controllers/AssemblyOperationsController.php
@@ -189,13 +189,13 @@ class AssemblyOperationsController extends Controller
     public function delete_assembly_operations($id)
     {
         // Try to find the client by the given ID
-        $get_assembly_operations_id = AssemblyOperationModel::select('assembly_operations_id')
+        $get_assembly_operations_id = AssemblyOperationModel::select('assembly_operations_id', 'company_id')
                                         ->where('id', $id)
                                         ->first();
         
         // Check if the client exists
 
-        if ($get_assembly_operations_id) 
+        if ($get_assembly_operations_id && $get_assembly_operations_id->company_id === Auth::user()->company_id) 
         {
             // Delete the client
             $delete_assembly_operations = AssemblyOperationModel::where('id', $id)->delete();
